+++ src/Ionization/Ionization.h
@@ -16,7 +16,7 @@ using namespace std;
 //! Class Ionization: generic class allowing to define Ionization physics
 class Ionization
 {
-   public:
+public:
     //! Constructor for Ionization
     Ionization(Params &params, Species *species);
     virtual ~Ionization();
@@ -24,25 +24,16 @@ class Ionization
     //! Overloading of () operator
     virtual void operator()(Particles *, unsigned int, unsigned int, std::vector<double> *, Patch *, Projector *, int = 0) {};
     //! method for envelope ionization
-    virtual void envelopeIonization(Particles *, unsigned int, unsigned int, std::vector<double> *, std::vector<double> *,
-                                    std::vector<double> *, std::vector<double> *, Patch *, Projector *, int = 0, int = 0) {};
-
-    // method for tunnel ionization using tasks
-    virtual void ionizationTunnelWithTasks(Particles *, unsigned int, unsigned int, std::vector<double> *, Patch *,
-                                           Projector *, int, int, double *, double *, double *, int = 0) {};
-    // join the lists of electrons created through ionization when tasks are used
-    void joinNewElectrons(unsigned int Nbins);
+    virtual void envelopeIonization( Particles *, unsigned int, unsigned int, std::vector<double> *, std::vector<double> *, std::vector<double> *, std::vector<double> *, Patch *, Projector *, int = 0, int = 0 ){};
 
     Particles new_electrons;
-    Particles *new_electrons_per_bin;
-
+    
     //! Whether the initial charge (of the atom that was ionized) should be saved
     bool save_ion_charge_ = false;
     //! Temporarily contains the initial charge of the atom that was ionized
     std::vector<short> ion_charge_;
-    std::vector<std::vector<short> > ion_charge_per_bin_;
 
-   protected:
+protected:
     double eV_to_au;
     double au_to_mec2;
     double EC_to_au;
@@ -55,7 +46,9 @@ class Ionization
     unsigned int nDim_particle;
     double ionized_species_invmass;
 
-   private:
+private:
+
+
 };
 
 #endif